Deduplicate LabelAndInput/ made into component.
