+++ mobile/src/pages/Register.tsx
@@ -4,36 +4,7 @@ import { TextInput, Text, Pressable, View, ScrollView } from "react-native";
 import Button from "../components/Button";
 import Icon from "react-native-vector-icons/Ionicons";
 import { register } from "../helpers/mutate";
-
-interface Props {
-  value: string;
-  labelText: string;
-  placeholder: string;
-  secureTextEntry?: boolean;
-  onChangeText: (value: string) => void;
-}
-
-function LabelAndInput({ value, labelText, placeholder, secureTextEntry, onChangeText}: Props){
-
-  return (
-    <>
-      <Text className="font-bold mb-3 text-xl">
-        {labelText}
-      </Text>
-
-      <View className="rounded-lg bg-slate-300 w-80 h-12 mb-5 items-start justify-center">
-        <TextInput
-          className="ml-3 w-80"
-          value={value}
-          placeholder={placeholder}
-          placeholderTextColor="#000000"
-          secureTextEntry={secureTextEntry}
-          onChangeText={onChangeText}
-        />
-      </View>
-    </>
-  );
-}
+import LabelAndInput from "../components/LabelAndInput";
 
 export default function Register({ navigation }) {
   const [email, setEmail] = useState("");
